Determination Points - finished Weapon DP
Attributes, Skills, Weapons & Magic spells are done full round trip

still need to do Languages

+++ ACS/src/com/starfyre1/dataModel/SkillsRecord.java
@@ -451,21 +451,21 @@ public class SkillsRecord implements LevelListener, Savable {
 		  		Listener is:				Chance to Hear 		Range
 				Human, Non-thief with a save		20%			60'
 				Vs. Surprise of less than 75%
-
-
+		
+		
 				Human, Non-thief with a save		30%			60'
 				Vs. Surprise of more than 75%
-
+		
 				Elven, Half-Elven and Dwarrow		30% 		120'
-
+		
 				Other--(some monsters will have		20%			60'
 				high hearing abilities)
-
+		
 				Thief					Varies with lvl		60' *
 					*	Add 5’ per 10% above 100% to hear.
-
+		
 				Halve all chances when listening to a door.
-
+		
 		*/
 
 		boolean innate = mIsInnate[HEAR];
@@ -1248,7 +1248,7 @@ public class SkillsRecord implements LevelListener, Savable {
 		if (skill >= bonus) {
 			return 0;
 		}
-		return (int) Math.ceil((bonus - skill) / 4);
+		return (int) Math.ceil(((double) bonus - (double) skill) / 4);
 
 	}
